- move useNavigateToParticipants to src/hooks/navigation
